test: validate CI hang fix
Dummy PR to validate CI completes without hanging.
Applies the same fixes from PR #888.
